Added strip command, and fixed some little things in rollback/rollforward
I also wasted a lot of time trying to hunt down an error that turned
out to be due to a list I used to hold original state before an
operation in a test being a pointer to the mutated state in what was
being altered.  Damn references and mutable data structures.