Moving some ugly sqlite hacks into python code.
This fixes refpkg_intersection, which was broken probably because it was
relying on undefined behavior in sqlite. It should, from now on, work reliably
without using any similar hacks.

As a result, there is also now an --all-ranks flag, just so you can know all
ranks instead of just the most specific.